requirements.txt
about-time==4.2.1
absl-py==1.4.0
addict==2.4.0
aiohttp==3.10.8
aiosignal==1.3.1
alive-progress==3.1.5
antlr4-python3-runtime==4.9.3
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
babel==2.16.0
cachetools==5.5.0
calmsize==0.1.3
ccimport==0.3.7
cftime==1.6.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comet-ml==3.47.0
commonmark==0.9.1
configobj==5.0.9
crc32c==2.7.1
cumm-cu111==0.2.9
cupy-cuda111==12.3.0
cycler==0.12.1
cython==3.0.11
dataclasses==0.6
deepspeed==0.15.1
deprecation==2.1.0
diffusers==0.30.3
docker-pycreds==0.4.0
drjit==0.4.6
dulwich==0.22.1
easydict==1.13
einops==0.8.0
everett==3.1.0
fastrlock==0.8.2
filelock==3.16.1
fire==0.7.0
flatbuffers==24.3.25
flatten-dict==0.4.2
fonttools==4.54.1
freetype-py==2.5.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.6.0
gitdb==4.0.11
gitpython==3.1.43
google-auth==2.27.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grapheme==0.6.0
grpcio==1.64.1
h5py==3.11.0
hjson==3.1.0
huggingface-hub==0.24.7
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib-metadata==8.4.0
importlib-resources==6.4.5
iopath==0.1.10
jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonschema==4.23.0
jupyter-packaging==0.12.3
jupyter-server==2.14.2
jupyterlab==4.2.5
jupyterlab-server==2.27.3
keras
keras-preprocessing==1.1.2
kiwisolver==1.4.7
kornia==0.7.3
lark==1.2.2
libclang==18.1.1
llvmlite
loguru==0.7.2
markdown==3.7
matplotlib==3.7.1
matplotlib2tikz==0.7.6
meshio==5.3.5
mitsuba==3.5.2
mrcfile==1.5.3
multidict==6.1.0
multipledispatch==1.0.0
mypy-extensions==1.0.0
nbclassic==1.1.0
nbformat==5.10.4
nestargs==1.1.0
netcdf4==1.7.1.post2
networkx
ninja==1.11.1.1
notebook-shim==0.2.4
numba
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.10.0.84
openexr==3.3.0
opt-einsum==3.4.0
pandas
pathtools==0.1.2
pccm==0.3.4
pip==24.1.2
plyfile
portalocker==2.10.1
progressbar2==4.5.0
promise==2.3
protobuf==3.20.3
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pybind11==2.13.6
pydeprecate==0.3.2
pyglet==2.0.17
pykeops==2.2.3

pyopengl==3.1.7
pyparsing==3.1.4
pyquaternion==0.9.9
pyrr==0.10.3
pyrsistent==0.20.0
python-swiftclient==4.6.0
python-utils
pytorch-lightning==2.4.0
pytz==2024.2
pywavelets
pyyaml==6.0.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.8.1
rsa==4.9
ruamel-yaml==0.18.6
ruamel-yaml-clib==0.2.8
scikit-image
scikit-learn
scipy
seaborn
semantic-version==2.10.0
sentry-sdk==2.15.0
sharedarray==3.2.4
shortuuid==1.0.13
simple-parsing==0.1.6
simplejson==3.19.3
smmap==5.0.1
sniffio==1.3.1
tabulate==0.9.0
tensorboard
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardx==2.6.2.2
tensorflow-io-gcs-filesystem
termcolor==2.3.0
tf-estimator-nightly==2.16.0.dev2024012409
tflearn==0.5.0
tfrecord==1.14.5
threadpoolctl==3.5.0
tifffile
tikzplotlib==0.10.1
tomlkit==0.13.2
torchmetrics==1.4.2
tqdm==4.66.5
trimesh==4.4.9
typing-extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.2.3
wandb==0.18.3
webcolors==24.8.0
websocket-client==1.8.0
werkzeug==3.0.4
wrapt==1.16.0
wurlitzer==3.1.1
yacs==0.1.8
yarl==1.13.1
yaspin==3.1.0
